+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Added
 - `BitSetInterface::is_empty`.
 - All `BitSetInterface`s now implement `Eq`.
+- `BitSet` (without &) now implements op's too.
 
 ## 0.2.0
 ### Changed